Lots of work to get arrays to serialize nicely. So far they are working pretty well. Performance is looking pretty good as well. Next stop is going to be extending the same array goodness to support generic lists as well.

+++ lang/dotnet/src/Avro/BinaryEncoder.cs
@@ -170,7 +170,7 @@ namespace Avro
 
         }
 
-        public void SetItemCount(Stream Stream, int value)
+        public void SetItemCount(Stream Stream, long value)
         {
             if (value > 0)
             {
